fix(renderer): keep parent session icon under nested Claude/Codex
Nested child process and title signals no longer rebrand a tab while
launch ownership is still active; reclaim only after the parent exits.
Fixes #13341.
